Practice 2 code finished
The project has a new controller with the url "listar". If no item id is given, it shows the article 2 (as expected), if the id given can't be found it returns a null item (thus causing an error), and if the id it's larger than 2 characters, it throws an error showing that there's an error since the url pattern doesn't match the regex expected (and defined).